Change endpoints and move into app file
