+++ projects-start-data.js
@@ -3,4 +3,8 @@ module.exports = [
     name: "Bathroom walls",
     palettes: [1, 2, 3]
   },
+  {
+    name: 'Kid bedroom',
+    palette: 1
+  }
 ];
